requirements.txt
streamlit
ultralytics
folium
streamlit-folium
opencv-python-headless
pillow
numpy